+++ performance.h
@@ -7,11 +7,11 @@
 class ETL;
 
 struct PerformanceStatistics {
-	int32_t total_write_cycles;
-	int32_t actual_total_write_cycles;
-	int32_t time_cost_in_sec;
-	int32_t RAM_cost;
-	int32_t total_write_bytes;
+	long long total_write_cycles;
+	long long actual_total_write_cycles;
+	long long time_cost_in_sec;
+	long long RAM_cost;
+	long long total_write_bytes;
 
 	PerformanceStatistics() {
 		this->total_write_cycles	= 0;
@@ -26,12 +26,12 @@ class ETLPerformance {
     public:
 	ETLPerformance(ETL* etl) : etl_(etl) {
 	}
-	float	GetStandardDeviation();
-	float	GetOverheadRatio();
-	int32_t GetRAMCost();
-	float	GetWriteSpeed();
-	void	StartTimer();
-	void	PrintInfo();
+	float	  GetStandardDeviation();
+	float	  GetOverheadRatio();
+	long long GetRAMCost();
+	long long GetWriteSpeed();
+	void	  StartTimer();
+	void	  PrintInfo();
 
     private:
 	ETL*  etl_;
